Character can move! Scheduler system is implemented! Background workers for many movement methods; item and creature are implemented! Tool.GetSystemMilliseconds is added! DepotChest is implemented! AddItem to NetworkMessage is implemented! Some coding improvements!

+++ GameServer/Model/ItemTemplate.cs
@@ -23,7 +23,7 @@ namespace HerhangiOT.GameServer.Model
         public Abilities Abilities { get; private set; }
         //ConditionDamage
 
-        public double Weight = 0.0;
+        public uint Weight = 0;
         public uint LevelDoor;
         public uint DecayTime;
         public uint WieldInfo;
